# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VolunteerConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /volunteers/profile:
    get:
      operationId: getVolunteerProfile
      tags: [volunteers]
      summary: Get own volunteer profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Volunteer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerProfile"
    put:
      operationId: updateVolunteerProfile
      tags: [volunteers]
      summary: Update volunteer profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVolunteerProfileRequest"
      responses:
        "200":
          description: Updated volunteer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerProfile"
...
  /volunteers/stats:
    get:
      operationId: getVolunteerStats
      tags: [volunteers]
      summary: Get volunteer impact stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Volunteer stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerStats"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateVolunteerProfileRequest:
      type: object
      properties:
        name:
          type: string
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        location:
          type: string
        lat:
          type: number
        lng:
          type: number
        bio:
          type: string
        availability:
          type: string
    VolunteerProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        location:
          type: string
        lat:
          type: number
        lng:
          type: number
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        bio:
          type: string
        availability:
          type: string
        totalHours:
          type: number
        eventsParticipated:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
    VolunteerStats:
      type: object
      properties:
        totalHours:
          type: number
        eventsParticipated:
          type: integer
        categoriesWorked:
          type: array
          items:
            type: string
        upcomingEvents:
          type: integer
      required:
        - totalHours
        - eventsParticipated
        - categoriesWorked
        - upcomingEvents
...